android: logger: Allow a UID to read it's own log entries
Modify the kernel logger to record the UID associated with
the log entries. Always allow the same UID which generated a
log message to read the log message.

Allow anyone in the logs group, or anyone with CAP_SYSLOG, to
read all log entries.

In addition, allow the client to upgrade log formats, so they
can get additional information from the kernel.

Change-Id: I023c91828557f4ca6f012e8c61481fcba1e77045
Signed-off-by: Nick Kralevich <nnk@google.com>
Reviewed-on: http://mcrd1-5.corpnet.asus/code-review/master/48196
Tested-by: Yeh Paris <paris_yeh@asus.com>
Reviewed-by: Sam hblee <Sam_hblee@asus.com>