PM: debug: Enable CONFIG_PM_DEBUG, debugging mask, and messages.
Kernel is compiled with CONFIG_PM_DEBUG set in the Power Management
code. This is helpful when debugging and reporting PM bugs, like
suspend support. Then, there is the file /sys/power/pm_test that
can be used to make the suspend core run in a test mode.

Enable debug message for checking wakelock status and android early
suspend and late resume routines.

Change-Id: I0bc75f2c217e2b7fc593bb1247b2335d0ec25686
Reviewed-on: http://mcrd1-5.corpnet.asus/code-review/master/45505
Reviewed-by: Yeh Paris <paris_yeh@asus.com>
Tested-by: Yeh Paris <paris_yeh@asus.com>
Reviewed-by: Wu Lucien <lucien_wu@asus.com>
Reviewed-by: Andy2 Kuo <andy2_kuo@asus.com>
Reviewed-by: Sam hblee <Sam_hblee@asus.com>