Sensors: A sysfs for checking compass compensation of factory calibrated gain.
In order to check the compass raw data has been multiplied by gain
that calibrated in factory, we create a new sysfs function to check it.

We cache both the data before/after the gain calculation
and ensure the data has been calibrated before data output.

It would not affect the performance about e-compass.
It's just a way for double-checking the real case when testing.

ps. # cat ../iio:device1/compass_cali_test could test this cmd.

Change-Id: I41e44252b58ac2e75c451b17314f9d60b40dec4c