tegra3: Add cheaper auto hotplug driver Instead of the mess which NVIDIA ships with their kernel, this driver is simple and cheap and aims to avoid unnecessary hotplug events and unnecessary task migrations, both of which the default NVIDIA mess does very poorly. Did I mention both are very expensive? :( The design is simple. The value of nr_running() (which is the current number of runnable threads) is averaged over 10 iterations by a ringbuffer. If the value is over a dynamically calculated value, all CPUs are onlined. When the value falls under another dynamically calculated value, the CPUs are offlined with a 2 second delay incase of an increase in load in the meantime.
If there is persistent load on the system, the CPUs will not continually thrash
between each other and migrate tasks like the stock driver. The number of online
CPUs will remain constant, until the load has dropped below the dynamically
calculated value.

The online and offline values are dynamically scaled based on the number of
CPUs online, as is the sampling rate of the driver. This aims to reduce the
expense of the driver and also reduce the number of unless hotplug transitions
which could have been avoided had cpufreq been given the chance to ramp the
online CPUs up and deal with the load.

This driver also ignores the LP core, as my recent tests have shown that there
is less battery drain when the LP core is disabled and LP2 is enabled (C2 state).

This driver will also offline secondary cores when the screen is turned off.
It also has a module param to set the minimum number of online CPUs:
/sys/module/tegra3_hotplug/parameters/min_online_cpus
Valid values are 1 to 4 inclusive, unless your tegra3 suddenly grew 64 cores.
This value is ignored when the device enters earlysuspend, as the secondary
CPUs are all offlined, but is set again when the screen is turned back on.